django/requirements/requirements.txt
Django==2.2.10
django_environ==0.4.5
psycopg2-binary==2.8.5
celery==4.4.2
